+++ VTK/Hybrid/vtk3DWidget.cxx
@@ -132,7 +132,7 @@ float vtk3DWidget::SizeHandles(float factor)
     {
     double radius, z;
     double windowLowerLeft[4], windowUpperRight[4];
-    float *viewport = renderer->GetViewport();
+    double *viewport = renderer->GetViewport();
     int *winSize = renderer->GetRenderWindow()->GetSize();
     double focalPoint[4];
