+++ VTK/Rendering/vtkSelectVisiblePoints.cxx
@@ -57,7 +57,7 @@ void vtkSelectVisiblePoints::Execute()
   vtkPointData *outPD=output->GetPointData();
   vtkIdType numPts=input->GetNumberOfPoints();
   float x[4];
-  float dx[3], z, diff;
+  double dx[3], z, diff;
   int selection[4];
   
   if ( this->Renderer == NULL )
